requirements.txt
streamlit==1.42.0
together==1.4.1
matplotlib==3.10.0
plotly==6.0.0
langchain-community==0.3.17
langchain-openai==0.3.5